Add reputation and signer selection

+++ dkg-primitives/src/types.rs
@@ -109,9 +109,9 @@ pub struct DKGPublicKeyMessage {
 	pub signature: Vec<u8>,
 }
 
-pub const KEYGEN_TIMEOUT: u128 = 10;
-pub const OFFLINE_TIMEOUT: u128 = 10;
-pub const SIGN_TIMEOUT: u128 = 3;
+pub const KEYGEN_TIMEOUT: u32 = 10;
+pub const OFFLINE_TIMEOUT: u32 = 10;
+pub const SIGN_TIMEOUT: u32 = 3;
 
 pub trait DKGRoundsSM<Payload, Output, Clock> {
 	fn proceed(&mut self, _at: Clock) -> Result<bool, DKGError> {
